Update serverside_table.py
Hello Sergio,

Really nice script! And it has been so useful for me.

I would like to suggest you one change on line 111, when you check for (anything)>0, it gives an exception because type(anything) != int, therefore, the rightone would be int(anything) >0;

Thanks a lot for the script!

cheers!